requirements.txt
requests==2.25.1
psycopg2==2.8.6
apache-airflow==1.10.9